started class hierarchy for measurand forecast texts
