+++ lib/wrf_forecast/text.rb
@@ -1,0 +1,12 @@
+#!/usr/bin/ruby
+# @Author: Benjamin Held
+# @Date:   2020-05-06 18:11:59
+# @Last Modified by:   Benjamin Held
+# @Last Modified time: 2020-05-06 18:18:19
+
+module WrfForecast
+
+  require "wrf_forecast/forecast/measurand_text"
+  require "wrf_forecast/forecast/english/temperature_text"
+
+end
